requirements.txt
google-genai
fastapi
uvicorn
pydantic
python-dotenv
Pillow
